Assignment 27: COMPLETE

+++ app/controllers/posts_controller.rb
@@ -46,6 +46,12 @@ class PostsController < ApplicationController
   def destroy
     @post = Post.find(params[:id])
 
+    if current_user.moderator?
+      flash[:alert] = "You must be an admin to do that."
+      redirect_to [@post.topic, @post]
+      return
+    end
+
     if @post.destroy
       flash[:notice] = "\"#{@post.title}\" was deleted successfully!"
       redirect_to @post.topic
@@ -67,7 +73,7 @@ class PostsController < ApplicationController
   def authorize_user
     post = Post.find(params[:id])
 
-    unless current_user == post.user || current_user.admin?
+    unless current_user == post.user || current_user.admin? || current_user.moderator?
       flash[:alert] = "You must be an admin to do that."
       redirect_to [post.topic, post]
     end
